Problem: com.eventsourcing is too polluted
Solution: move repository related stuff to com.eventsourcing.repository
