+++ eventsourcing-core/src/main/java/com/eventsourcing/index/CQIndexEngine.java
@@ -8,6 +8,8 @@
 package com.eventsourcing.index;
 
 import com.eventsourcing.*;
+import com.eventsourcing.repository.Journal;
+import com.eventsourcing.Repository;
 import com.googlecode.cqengine.ConcurrentIndexedCollection;
 import com.googlecode.cqengine.IndexedCollection;
