+++ eventsourcing-core/src/main/java/com/eventsourcing/repository/BundleEventSetProvider.java
@@ -5,10 +5,9 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.eventsourcing.osgi;
+package com.eventsourcing.repository;
 
 import com.eventsourcing.Event;
-import com.eventsourcing.EventSetProvider;
 
 import java.util.HashSet;
 import java.util.List;
