+++ eventsourcing-core/src/main/java/com/eventsourcing/repository/ClassEntitySubscriber.java
@@ -5,7 +5,9 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.eventsourcing;
+package com.eventsourcing.repository;
+
+import com.eventsourcing.Entity;
 
 /**
  * {@link EntitySubscriber} that subscribes to all entities that are inherited from
