+++ eventsourcing-core/src/main/java/com/eventsourcing/repository/DispatchingCommandConsumer.java
@@ -5,8 +5,9 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.eventsourcing;
+package com.eventsourcing.repository;
 
+import com.eventsourcing.Command;
 import com.eventsourcing.hlc.HybridTimestamp;
 import com.eventsourcing.hlc.PhysicalTimeProvider;
 import com.eventsourcing.index.IndexEngine;
@@ -21,7 +22,7 @@ import java.util.*;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ForkJoinPool;
 
-public class DispatchingCommandConsumer extends AbstractService implements CommandConsumer {
+class DispatchingCommandConsumer extends AbstractService implements CommandConsumer {
 
     private List<CommandConsumer> consumers;
     private ServiceManager serviceManager;
