+++ eventsourcing-core/src/main/java/com/eventsourcing/repository/EntitySubscriber.java
@@ -5,7 +5,11 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.eventsourcing;
+package com.eventsourcing.repository;
+
+import com.eventsourcing.Entity;
+import com.eventsourcing.EntityHandle;
+import com.eventsourcing.Repository;
 
 import java.util.stream.Stream;
