+++ eventsourcing-core/src/main/java/com/eventsourcing/repository/JournalEntityHandle.java
@@ -5,8 +5,10 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.eventsourcing;
+package com.eventsourcing.repository;
 
+import com.eventsourcing.Entity;
+import com.eventsourcing.EntityHandle;
 import lombok.Getter;
 import lombok.experimental.Accessors;
