+++ eventsourcing-core/src/main/java/com/eventsourcing/repository/MemoryLockProvider.java
@@ -5,8 +5,9 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.eventsourcing;
+package com.eventsourcing.repository;
 
+import com.eventsourcing.Lock;
 import com.google.common.util.concurrent.AbstractService;
 import org.osgi.service.component.annotations.Component;
