+++ eventsourcing-core/src/main/java/com/eventsourcing/repository/RepositoryMBean.java
@@ -5,10 +5,12 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.eventsourcing;
+package com.eventsourcing.repository;
 
-import java.util.Set;
+public interface RepositoryMBean {
+    boolean isRunning();
 
-public interface EventSetProvider {
-    Set<Class<? extends Event>> getEvents();
+    String[] getInstalledCommands();
+
+    String[] getInstalledEvents();
 }
