+++ eventsourcing-h2/src/main/java/com/eventsourcing/h2/MVStoreJournal.java
@@ -13,6 +13,12 @@ import com.eventsourcing.hlc.HybridTimestamp;
 import com.eventsourcing.layout.Deserializer;
 import com.eventsourcing.layout.Layout;
 import com.eventsourcing.layout.Serializer;
+import com.eventsourcing.repository.Journal;
+import com.eventsourcing.repository.JournalEntityHandle;
+import com.eventsourcing.repository.JournalMBean;
+import com.eventsourcing.Repository;
+import com.eventsourcing.repository.LockProvider;
+import com.eventsourcing.utils.CloseableWrappingIterator;
 import com.google.common.base.Joiner;
 import com.google.common.collect.Iterators;
 import com.google.common.io.BaseEncoding;
