+++ src/jmh/java/com/eventsourcing/jmh/MVStoreJournalBenchmark.java
@@ -7,9 +7,8 @@
  */
 package com.eventsourcing.jmh;
 
-import com.eventsourcing.Journal;
+import com.eventsourcing.repository.Journal;
 import com.eventsourcing.h2.MVStoreJournal;
-import com.eventsourcing.index.IndexEngine;
 import org.h2.mvstore.MVStore;
 
 import java.io.File;
